pyproject.toml
[project]
name = "gairdenier"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic-ai>=1.0.15",
    "python-dotenv>=1.1.1",
]

[dependency-groups]
dev = [
    "pre-commit>=4.3.0",
    "ruff>=0.13.3",
]
